create messages.py
