+++ qse/calc/pulser.py
@@ -4,17 +4,8 @@ the ASE's CP2K calculator, though at the end it may look very different from
 ASE calculator.
 https://pulser.readthedocs.io/en/stable/
 """
-
-import os
-import os.path
-from abc import ABCMeta
-from subprocess import PIPE, Popen
-from warnings import warn
-
-import ase.io
 import numpy as np
 import pulser
-import pulser.pulse
 import pulser.waveforms
 
 # from pulser_simulation import Simulation, SimConfig, QutipEmulator
@@ -24,8 +15,9 @@ from qse.calc import signal
 from qse.calc.calculator import (
     Calculator,
     CalculatorSetupError,
-    Parameters,
-    all_changes,
+)
+from qse.calc.messages import (
+    CalculatorSetupError,
 )
 
 # from ase.calculators.calculator import (Calculator, all_changes, Parameters, CalculatorSetupError)
